requirements.txt
cryptography
python-nmap
scapy
